x/x11 - Port missing makepkg/ script fix to x11.SlackBuild
e.g. If you try and bump BUILD on xorg-server, it won't get picked up,
since x11.SlackBuild only updates one of two variables if a file exists,
either MODBUILD (in the case of xorg-server) or SRCDIRBUILD.

(The version I sent to Pat fixed a second occurence of this - so fix that
in Slamd64 now that Pat has accepted the fix).