x/x11 - Rebuild libXft
Rebuilding libXft fixes bold rendering issues with Chinese fonts
(tested on Slamd64 as well, after discussions with Bruce Hill). We're
not sure what it is about the rebuilding that fixes the fonts, but
my guess is because we're rebuilding against a much newer Freetype
than when libXft was originally compiled.