c/xlibs32 - Add SlackBuild for xlibs32
Add a new SlackBuild to automatically build xlibs32. This replaces most
of the old copy + pasted SlackBuild's and configure files for the X
libraries, and condenses them into a handful of files.

This script also relies on their being a symlink, 'src', pointing to
source/x/x11/src/lib.

The file 'modularize' controls which libraries are built, and in what
order, so it is trivial to add or remove other libraries from xlibs32
in future.