l/ncurses - Fix bug 221
Fix the corruption in other applications (e.g. the nVidia installer) by
setting the necessary configure flag.

This fixes bug 221.

(This is an old fix, and the updated package is already in Slamd64 12.0.
However, the update to the SlackBuild was not included, so lets add this
in to keep a proper history, before upgrading ncurses again).